refactored type definitions
