+++ src/components/ui/exercises/lesson-navigation.tsx
