+++ src/components/ui/lesson/index.tsx
@@ -1,5 +1,5 @@
 export { default as LessonPlayer } from './lesson-player';
-export { default as LessonNavigation } from './lesson-navigation';
+export { default as LessonNavigation } from '../exercises/lesson-navigation';
 export { default as LessonProgressBar } from './lesson-progress-bar';
 export { default as PageTemplate } from './page-template';
 export { default as ContentRenderer } from './content-renderer';
