+++ src/components/ui/lesson/page-template.tsx
@@ -17,8 +17,7 @@ export const PageTemplate: React.FC<PageTemplateProps> = ({ page }) => {
       animate={{ opacity: 1, x: 0 }}
       exit={{ opacity: 0, x: -20 }}
       transition={{ duration: 0.3 }}
-      className="space-y-6" // Predetermined spacing between content items
-    >
+      className="space-y-6">
       {page.title && <h2 className="text-xl font-serif text-roman-red mb-4">{page.title}</h2>}
 
       {page.items.map((item, index) => (
@@ -27,8 +26,7 @@ export const PageTemplate: React.FC<PageTemplateProps> = ({ page }) => {
           initial={{ opacity: 0, y: 20 }}
           animate={{ opacity: 1, y: 0 }}
           transition={{ duration: 0.3, delay: index * 0.1 }}
-          className="space-y-4" // Additional spacing within each content item
-        >
+          className="space-y-4">
           <ContentRenderer content={item} />
         </motion.div>
       ))}
