+++ src/types/exercise.d.ts
@@ -1,0 +1,9 @@
+export type {
+  BaseExercise,
+  MatchingExercise,
+  FillExercise,
+  TextSelectionExercise,
+  VerbAnalysisExercise,
+  VerbConjugationExercise,
+  Exercise,
+} from './exercises';
