+++ src/types/exercises/base.d.ts
@@ -1,0 +1,5 @@
+import { ContentItem } from '../content';
+
+export interface BaseExercise extends ContentItem {
+  instructions: string;
+}
