+++ src/types/exercises/fill.d.ts
@@ -1,0 +1,12 @@
+import { BaseExercise } from './base';
+
+export interface FillExercise extends BaseExercise {
+  type: 'fill';
+  data: {
+    items: {
+      text: string;
+      answer: string;
+      hint?: string;
+    }[];
+  };
+}
