+++ src/types/exercises/index.d.ts
@@ -1,0 +1,19 @@
+export type { BaseExercise } from './base';
+export type { MatchingExercise } from './matching';
+export type { FillExercise } from './fill';
+export type { TextSelectionExercise } from './text-selection';
+export type { VerbAnalysisExercise } from './verb-analysis';
+export type { VerbConjugationExercise } from './verb-conjugation';
+
+import type { MatchingExercise } from './matching';
+import type { FillExercise } from './fill';
+import type { TextSelectionExercise } from './text-selection';
+import type { VerbAnalysisExercise } from './verb-analysis';
+import type { VerbConjugationExercise } from './verb-conjugation';
+
+export type Exercise =
+  | MatchingExercise
+  | FillExercise
+  | TextSelectionExercise
+  | VerbAnalysisExercise
+  | VerbConjugationExercise;
