+++ src/types/exercises/text-selection.d.ts
@@ -1,0 +1,14 @@
+import { BaseExercise } from './base';
+
+export interface TextSelectionExercise extends BaseExercise {
+  type: 'text-selection';
+  data: {
+    passage: string;
+    questions: {
+      id: string;
+      text: string;
+      correctWord: string;
+      explanation?: string;
+    }[];
+  };
+}
