+++ src/types/index.d.ts
@@ -1,0 +1,17 @@
+export type { ContentItem, TextContent, EmphasisContent, TableContent, ComponentNarration } from './content';
+
+export type { VocabularyItem, VocabularyContent } from './vocabulary';
+
+export type {
+  BaseExercise,
+  MatchingExercise,
+  FillExercise,
+  TextSelectionExercise,
+  VerbAnalysisExercise,
+  VerbConjugationExercise,
+  Exercise,
+} from './exercises';
+
+export type { RenderableContentItem, BasePage, IntroductionPage, ExercisePage } from './page';
+
+export type { Lesson } from './lesson';
